Consolidating modules into logic container
I had kept the in memory modules and classes separate as part of the
initial development. However, as I was looking at the code, I opted to
factor them into the `InMemoryAdapater`. Prior to the adapter
implementation the modules that were moved into InMemoryAdapter were
referenced directly throughout the feature specs and even in the
production implementation.

However, with the clear boundary defined in the AbstractAdapter,
and the implementation in the InMemoryAdapter, the Indexer code no
longer needs direct reference those concepts.