Adding license declaration to spec
Leveraging preferred mechanism for language declaration:

See - https://spdx.org/licenses/

We are looking at leveraging [Pivotal Lab's LicenseFinder](https://github.com/pivotal/LicenseFinder#a-plea-to-package-authors-and-maintainers)
for reviewing the licenses of our dependencies. Rubydora's returns an
unknown for its license.

(It scans the gemspec and found nothing, it also compared the
LICENSE.txt against the known templates and they were not exact
matches)

[skip ci]