Adding additional examples and clarifications
I'm including notations for membership to ensure a common understanding
of language. Then adding a few more tests that highlight updating a
document into a cyclical state.

This does not resolve the larger problem of relying on the time to live
to kill the indexing process instead of checking within that time to
live for a cyclical graph.