Adding detection for indexing self ancestry
Prior to this commit, we relied on the time to live enforcing cycles.
The reality, however, is that if we had the following:

> A ={ B ={ C ={ B

We would be updating the index document multiple times and adding a
pathname of the following form: "A/B/C/B/C/B/C". This code short-
circuits that and hopefully keeps the index in a better state.

Beware the Delta Wave http://futurama.wikia.com/wiki/Delta_Brainwave