Adding #deepest_nested_depth for indexing doc
In working through Samvera nesting indexing, it became clear that it
would be helpful to know the current depth of an indexed document. This
is something easily calculated when the
`Samvera::NestingIndexer::Document::IndexDocument` is being created.

As part of this change, I pushed code to use the already existing
`Samvera::NestingIndexer::Document::IndexDocument`; which removes some
duplication of knowledge and hints at future steps to remove
duplication.

A future change would be to shift from primitive parameters when writing
to the index layer to a
`Samvera::NestingIndexer::Document::IndexDocument`. In this way future
changes to the API would be a bit less brittle (e.g. wouldn't have to
add new keywords to the parameter list).