Replacing multi-parameters with parameter object
It is good practice to question any method with more than 3 parameters.
I believe this is especially important when the method defines an
interface to external implementations.

This commit preserves the existing method, though it no longer uses
that method. The gem now uses the new method.

Implementors will need to replace their existing
`.write_document_attributes_to_index_layer` method with a
slight modification `.write_nesting_document_to_index_layer`.

This is a breaking upstream change. However, I believe there are few
implementations of this, and as such is something we can change if we
coordinate updates to upstream changes:

* https://github.com/samvera/hyrax/blob/collections-sprint/app/services/hyrax/adapters/nesting_index_adapter.rb#L61-L81