Tightening up public interface definition
This includes an explicit definition of what is in the public API, as
well as what is slated to be removed from the public API.

My goal, in this effort, is to begin to explore what semblance of
automation we might be able to perform. I think it is beyond feasible
at this point, but I want to think about how we, the community, are
modeling the declaration of a public interface.