Memoize initial_index_document
Save value to avoid multiple calls to adapter to load the same
information.

Also add optional "extent" parameter to `reindex_relationships` method,
and pass it through the adapter call to `each_child_document_of`. When
used, the adapter receives the value and can vary behavior based it. The
intent of this method is to allow adapter to determine which children
to return for reindexing rather than to reindex the full graph.